+++ portal/webpack.config.js
@@ -5,13 +5,14 @@ module.exports = {
     'entry': path.resolve(__dirname, 'src', 'index.js'),
     'output': {
         'filename': 'main.js',
-        'publicPath': '/static/'
+        'library': 'myeqip',
+        'libraryTarget': 'window'
     },
     'module': {
         'rules': [
             {
                 'test': /\.js$/,
-                'exclude': [ /node_modules/, /static_serve/ ],
+                'exclude': [ /node_modules/ ],
                 'use': {
                     'loader': 'babel-loader',
                     'options': {
@@ -21,7 +22,7 @@ module.exports = {
             },
             {
                 'test': /\.scss$/,
-                'exclude': [ /node_modules/, /static_serve/ ],
+                'exclude': [ /node_modules/ ],
                 'use': [
                     {
                         'loader': 'file-loader',
@@ -34,12 +35,14 @@ module.exports = {
                     { 'loader': 'sass-loader' }
                 ]
             },
+            {
+                'test': /\.svg$/,
+                'exclude': [ /node_modules/ ],
+                'use': {
+                    'loader': 'svg-url-loader'
+                }
+            }
         ]
-    },
-    'resolve': {
-        'alias': {
-            'static': path.resolve('static'),
-        }
     },
     'plugins': [
         new CleanWebpackPlugin()