+++ CHANGELOG.md
@@ -1,14 +1,32 @@
 # CHANGELOG
 
+## 0.0.2
+
+### Interface
+
+- [Footer content](https://github.com/catalpainternational/myeqip-portal/pull/31)
+- [RichTextBlock](https://github.com/catalpainternational/myeqip-portal/pull/35)
+- [Splits menu into two `nav`s.](https://github.com/catalpainternational/myeqip-portal/pull/38)
+- [QuoteBlock](https://github.com/catalpainternational/myeqip-portal/pull/40)
+- [PostcardBlock](https://github.com/catalpainternational/myeqip-portal/pull/42)
+- [Top and side navigation](https://github.com/catalpainternational/myeqip-portal/pull/46)
+- [Style rich text block](https://github.com/catalpainternational/myeqip-portal/pull/49)
+- [Style quote block](https://github.com/catalpainternational/myeqip-portal/pull/52)
+- [Style title block](https://github.com/catalpainternational/myeqip-portal/pull/57)
+
+### QA
+
+### Code
+
 ## 0.0.1
 
 ### Interface
 
-
+- [Support for multiple languages](https://github.com/catalpainternational/myeqip-portal/pull/15)
+- [Log in and log out support](https://github.com/catalpainternational/myeqip-portal/pull/17)
+- [Search bar](https://github.com/catalpainternational/myeqip-portal/pull/21)
 
 ### QA
-
-
 
 ### Code